+++ gulpfile.babel.js
@@ -10,6 +10,7 @@ import { buildJs, buildPolyfills } from './build/scripts'
 import { buildCss } from './build/styles'
 import { serve } from './build/serve'
 import { watch } from './build/watch'
+import { buildSvgSprite } from './build/icons'
 
 // modulus `create:*`
 import createPlugin from './build/modulus/create-plugin'
@@ -24,6 +25,7 @@ gulp.task('build:css', () => buildCss(...cfg.src.scss.entries.map(entry => entry
 gulp.task('build:html', () => buildPanini())
 gulp.task('build:js', () => buildJs(...cfg.src.js.entries.map(entry => entry.file)))
 gulp.task('build:polyfills', () => buildPolyfills(...cfg.src.polyfills.entries.map(entry => entry.file)))
+gulp.task('build:icons', () => buildSvgSprite())
 gulp.task('serve', (done) => serve(done))
 gulp.task('watch', (done) => watch(done))
 
@@ -33,5 +35,5 @@ gulp.task('create:controller', (done) => createController(__dirname, yargs.argv.
 gulp.task('create:plugin', (done) => createPlugin(__dirname, yargs.argv.name, done))
 
 // npm global tasks
-gulp.task('build', gulp.series('clean:html', 'build:html', 'build:css', 'build:js', 'build:polyfills', 'copy:assets'))
+gulp.task('build', gulp.series('clean:html', 'build:icons', 'build:html', 'build:css', 'build:js', 'build:polyfills', 'copy:assets'))
 gulp.task('default', gulp.series('build', 'serve', 'watch'))